Add Dart VM container assertions for valgrind under cargo test
