+++ tests/vm.rs
